+++ src/main/java/com/solarlab/adboard/controller/ImageController.java
@@ -1,5 +1,6 @@
 package com.solarlab.adboard.controller;
 
+import com.solarlab.adboard.config.SecurityUtils;
 import com.solarlab.adboard.dto.response.ExceptionResponse;
 import com.solarlab.adboard.dto.response.image.ImageResponse;
 import com.solarlab.adboard.mapper.ImageMapper;
@@ -14,7 +15,6 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.constraints.PositiveOrZero;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
@@ -30,6 +30,7 @@ public class ImageController {
 
     private final ImageService imageService;
     private final ImageMapper imageMapper;
+    private final SecurityUtils securityUtils;
 
     @Operation(summary = "Get advertisement images")
     @ApiResponses({
@@ -60,11 +61,11 @@ public class ImageController {
                     content = @Content(schema = @Schema(implementation = ExceptionResponse.class)))
     })
     @PostMapping("/upload")
-    @PreAuthorize("@securityUtils.isAdvertisementOwner(#advertisementId)")
     public ResponseEntity<ImageResponse> uploadImage(
             @PositiveOrZero @PathVariable(name = "advertisementId") Long advertisementId,
             @RequestParam("file") MultipartFile file
     ) {
+        securityUtils.ensureAdvertisementOwner(advertisementId);
         Image image = imageService.uploadImage(file, advertisementId);
         return ResponseEntity.ok(imageMapper.toImageResponse(image));
     }
@@ -81,11 +82,11 @@ public class ImageController {
             @ApiResponse(responseCode = "404", description = "Image not found",
                     content = @Content(schema = @Schema(implementation = ExceptionResponse.class)))
     })
-    @PreAuthorize("@securityUtils.isImageOwner(#id)")
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteImage(
             @PositiveOrZero @PathVariable(name = "id") Long id
     ) {
+        securityUtils.ensureImageOwner(id);
         imageService.deleteImage(id);
         return ResponseEntity.noContent().build();
     }
